+++ src/js/engagement_view/src/components/graphDisplay/utils/routing.tsx
@@ -1,4 +1,4 @@
-import { RouteState } from './GraphVizCustomTypes';
+import { RouteState } from '../../../types/CustomTypes';
 
 
 export const defaultRouteState = (): RouteState => {
